requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
dateparser==0.7.2
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pyfiglet==0.8.post1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
six==1.14.0
termcolor==1.1.0
tzlocal==2.0.0
urllib3==1.25.8
